Add Shared Terraform Workflows
Why these changes are being introduced:
InfraEng has many (many!) Terraform repos, all with the same three
workflows: validate Terraform, run checkov security check, update
README with terraform-docs output. The "validate Terraform" is linked
to a specific version of Terraform, and if that ever had to be updated
it would be a total nightmare. Seems like a good use case for a set
of shared workflows.

This needs to be coordinated with updates to the
mitlib-tf-template repository. There is currently a feature branch
in that repo with changes to the workflows to switch from the three
workflows to the since "caller" workflow. Once this branch is committed
to `main`, that feature branch will need to be updated to reference
the `main` branch for the shared workflows.

How this addresses that need:
* Add three Terraform-related workflows to the shared workflows list
* Update the README to reflect the new workflows

Side effects of this change:
None.