WIP Lambda Deployment Shared Workflow
Why these changes are being introduced:
The POD/ReShare project has two apps that will be containers wrapped
in a Lambbda function. This is going to be a typical deployment
scenario for us with Step Functions, so it seems wise to build a shared
workflow for the automated deployment of this objects. This is just a
work-in-progress for now and is designed for only the dev environment.

How this addresses that need:
* Create new shared workflow for publishing Lambda-wrapped containers
* Update README with details

Limitations:
* This is currently designed to only work in the dev (Dev1)
environment
* It is heavily dependent on consistent Makefiles in our application
repos (it simply calls `make` commands)
* It requires three values to be passed to the shared workflow (details
in the README)
* It is only designed for containers that are intended to be wrapped
in Lambda functions (the shared workflow for Fargate containers will
be a future commit)