Update Ruby CI to run on Ubuntu 22.04
Why these changes are being introduced:

We are upgrading our Rails apps to use the Heroku 22 stack, which
requires Ubuntu 22.04.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-183

How this addresses that need:

This explicitly declares ubuntu-22.04 as the os for test job,
rather than ubuntu-latest.

Side effects of this change:

The system setup block has been removed, as `libsqlite3-dev` is
included in the new Heroku stack.