Force Docker BuildKit On
Why these changes are being introduced:
When testing the automated build/deploy in Dev1 for docker-matomo,
GitHub Actions threw an error on the build command:
`the --chmod option requires BuildKit. Refer to \
https://docs.docker.com/go/buildkit/ to learn how to build images with \
BuildKit enabled`

This is the first attempt to address this problem.

How this addresses that need:
* Add `DOCKER_BUILDKIT=1 ` in front of the `docker build` command

Side effects of this change:
Unknown.