Enable BuildKit for Docker
During the testing of the docker-matomo automated builds, the dev-build
workflow generated an error during the `docker build` step. The error
indicated that the BuildKit needed to be enabled in order for it to
succeed. The [Docker Build Enhancements](https://docs.docker.com/develop/develop-images/build_enhancements/)
documentation indicated that the simplest way to enable the BuildKit
is to prefix the `docker build` command with `DOCKER_BUILDKIT=1`.

How this addresses that need:
* Update each of the shared workflows that uses the `docker build`
command by prefxing `DOCKER_BUILDKIT=1 `
* Update the README to correctly reflect the existing shared workflows
for our Fargate and Lambda deployment automation

Side effects of this change:
None. The previous commit tested this for the dev-build workflow and
the test run from docker-matomo was successful. This extends the fix
to all the other build automations.