Improve ECR Container Tagging
Why these changes are being introduced:
The current tags on containers that are pushed by the GitHub Actions
workflows get tags that are meaningless (because of the way the checkout
command works in GHA). The changes introduced here improve on the
container tagging so that the tags match up to commits in the
repository that calls these shared workflows.

In particular, in Dev1, the tags on a container pushed to an ECR
repository will get
* the short SHA of the most recent commit from the feature branch in
the PR
* the PR number (in the format "PR-#") of the pull request that
generated the updated container
* the word "latest"

In Stage-Workloads, the tags on the container in the ECR repo will be
* the short SHA of the merge commit in the `main` branch from the PR
* the word "latest"

In Prod-Workloads, the tags on the container will be
* the short SHA of the commit that was tagged as a release
* the release tag itself
* the word "latest"

How this addresses that need:
* Add logic to address both automated runs and manually triggered runs
of the dev-deploy workflow for fargate
* Use values from the GITHUB_CONTEXT in the runner to capture the
latest commit SHA from the feature branch in the PR for tagging (in the
dev-deploy workflows)
* Use values from the GITHUB_CONTEXT in the runner to capture the
PR number (in the dev-deploy workflows)
* Use values from the GITHUB_CONTEXT in the runner to capture the
merge commit SHA (in the stage-deploy workflows)
* Use values from the GITHUB_CONTEXT in the runner to capture the
release tag on `main` (in the prod-promote workflows)
* Update README with more details on the auto-deploy workflows