Simplify ECR Deploys
Why these changes are being introduced:
The carbon build process required a few extra steps before
`docker build` would run correctly. For automated deploys to dev & stage
and promotes to prod to work, the shared workflows needed to be able
to handle those additional steps. Additionally, there were duplicate
workflows for Fargate deploys and Lambda deploys that should be
consolidated for simpler (and DRYer) code. Finally, this repo should
have a PR template like all our other repos.

How this addresses that need:
* Create a single new ecr-shared-deploy-dev workflow file that will
handle both Fargate containers and Lambda containers/functions
* Create a single new ecr-shared-deploy-stage workflow file that will
handle both Fargate containers and Lambda containers/functions
* Create a single new ecr-shared-promote-prod workflow file that will
handle both Fargate containers and Lambda containers/functions
* Add checks for `.python-version` and `.ruby-version` files in the
repo (in the new dev and stage workflows)
* Clean up the setup python step using actions/setup-python only if
the `.python-version` file exists in the root of the repo (in the new
dev and stage workflows)
* Add a setup ruby step using ruby/setup-ruby only if the
`.ruby-version` file exists in the root of the repo (in the new dev and
stage workflows)
* Clean up the "pre-build" step so that it is universal (only depends
on the caller workflow specifying the commands)
* Switch to using $GITHUB_ENV from the deprecated set-output method
* Update the aws-credential action to use the Node.js 16 version
* Update in-line comments
* Update README
* Add a PR template file

Side effects of this change:
None. We leave the old workflows in place for now, until we can
update all the dependent repos.