Create Shared CDN Publishing Workflow
Why these changes are being introduced:
We will have multiple "sites" (e.g., GH repos) that will have static
content that needs to be published to our S3-based CDN. Since more than
one repo will need to run the automated publishing, it makes the most
sense to build the workflow here and call it from the other repos.

How this addresses that need:
* Create a shared workflow for syncing contents of a GitHub repo to
a folder in our S3-based CDN.
* Update README to document the required inputs for the new shared
workflow

Note: This is designed only for publishing to "stage" and "prod" CDNs.
And, it is designed (for now) to only ignore the .git and .github
directories.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/LM-93