Include CDN Cache Invalidation
Why these changes are being introduced:
After changes to objects in S3, it's safest to invalidate the cache
for the path that was updated. It's also a good idea to have the
ability to authenticate to our DEV environment for testing.

How this addresses that need:
* Add a step that uses the AWS CLI to invalidate the cache for the CDN
for the path in question
* Add DEV authentication block
* Cleanup README

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/LM-93