Add Option for Additional Ignored Folder
Why these changes are being introduced:
We need the ability to allow the web developer to optionally add an
additional folder to the "exclude" list in the aws cli sync command.
For example, this will allow for a documentation folder to be in the
static site repo but not allow the contents of that folder to be
synced to the website.

How this addresses that need:
* Add an optional workflow input ("EXCLUDE")
* Update the workflow sync step to use the additional EXCLUDE value if
it is passed to this shared workflow

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/LM-118
* https://mitlibraries.atlassian.net/browse/LM-93