Switch to Addtional Sync Parameters
Why these changes are being introduced:
The original attempt was restricted to adding a single excluded file
or folder. This was too limiting. This change addresses a concern
raised by the reviewer that we will likely need more flexibility than
just one file or folder.

How this addresses that need:
* Instead of a single `--exclude`, the input variable is configured as
a string to tack on the end of the `aws s3 sync` command.
* Update README for new input values for this workflow

Side effects:
Any pre-existing caller workflows will need to be updated to match
the new input value name.

Relevant tickets:
* https://mitlibraries.atlassian.net/browse/LM-93