Update for Multi Distribution CloudFront
Why these changes are being introduced:
We are in the process of moving to more than one CloudFront Distribution
related to our overall web content cleanup/migration (especially the
libraries.mit.edu static content). As a result, the publishing workflow
needs to be sensitive to the multiple distributions.

How this addresses that need:
* Add a DOMAIN input variable to capture the the standard CDN versus
the custom domain CDN
* Update the cache invalidation step to be aware of which cache is
getting invalidated
* Separate the `aws s3 sync` command into two different steps for the
two different CDNs (together with if statements so only one sync
actually runs)
* Add `echo` statments to update the $GITHUB_STEP_SUMMARY environment
variable to build a summary that will print to the UI in the Actions
tab in GitHub
* Update the README with additional details

Side effects of this change:
This will break existing caller workflows once it is merged to the
`main` branch. This cannot be deployed until after we have finished
deploying the updates from the mitlib-tf-workloads-libraries-website
repo through prod.