Remove pipenv cache from ECR workflows
Why these changes are being introduced:
There is a known bug in `actions/setup-python` related to caching:
https://github.com/actions/setup-python/issues/436
that we experienced with the first auto-deploy runs for the
alma-creditcardslips application. As a quick fix, we simply remove
the cache option for pipenv from the step and everything runs smoothly.

How this addresses that need:
* Remove `cache: 'pipenv'` from the actions/setup-python for both the
dev and stage ecr-shared-deploy- workflows.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-735