Publish Lambda@Edge Shared Workflow
Why these changes are being made:
As part of the GDT project we determined that we need a better way to
manage the deployment of Lambda@Edge functions in our CloudFront
distribution. Since we will have more than one application repository
building Lambda functions for Lambda@Edge, it made the most sense to
create a shared workflow here that can be called from multiple
application repositories. The overview of the workflow can be found
in Confluence here:
* https://mitlibraries.atlassian.net/l/cp/fJZhzHFq

How this addresses that need:
* Update the README with additional information about the new workflow
* Add a new workflow, cf-lambda-shared-deploy.yml, that runs all the
steps to push a .zip of the application code for the Lambda function
to the shared-files deployment bucket in AWS and then triggers a re-run
of the appropriate workspace in Terraform Cloud.

Side effects of this change:
None.

Related Jira tickets:
* https://mitlibraries.atlassian.net/browse/GDT-97