More Freedom for Packaging Zip File
Why these changes are being introduced:
Per comments in the PR conversation, some Lambda@Edge functions will
have more complicated packaging procedures than a simple `zip -j`
command. The updates here hand off the packaging and upload commands to
the application developer to be written in the caller repository's
Makefile as `create-zip` and `upload-zip` commands.

How this addresses that need:
* Remove unnecessary workflow inputs
* Set `make create-zip` and `make upload-zip` as the packaging commands
in the upload-zip job
* Update the TfC plan message to include the name of the caller
workflow for better auditing in TfC

Side effects of this change:
This introduces a hard dependency on the caller workflow have the
required `make` commands already defined. This will mean fixing
any repositories that are already using this workflow.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-97