Only Release Tags on Main Will Run
Why these changes are being introduced:
We need to ensure that only published release tags on the `main` branch
of the calling repo can run a production deploy.

How this addresses that need:
* Add a conditional to the `ecr-shared-promote-prod.yml` workflow job
to check if the call is coming from the `main` branch

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-991