Additional Logic for Prod Promote
Why these changes are being introduced:
In the last PR, a scenario was proposed that showed it was possible for
the wrong container to be deployed to Production. The changes here are
an attempt to prevent that scenario from happening by adding an
additional check to the prod-promote shared workflow. Thanks to Graham
for the initial seed for this solution. Thanks are also due to ChatGPT
for quickly providing the correct `sed` and `grep` piping to capture
the SHA tag.

How this addresses that need:
* Add a check to the ecr-shared-prod-promote workflow that compares the
8-character SHA of the HEAD commit on the `main` branch to the SHA tag
on the container in Stage-Workloads. If those two values match, the
workflow will continue running; if those two values don't match, the
workflow will fail with a notice to the GitHub Actions UI.

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-992