Adjust GITHUB_TOKEN Permissions
Why these changes are being introduced:
After testing changes in the Dev workflow, all three container
deployment workflows are updated. This is an interim step to resolve the
permissioning problem that started to crop up between the caller and
shared workflows. Other Terraform related workflows are updated too.

How this addresses that need:
* Update the `permissions:` to reflect on the necessary read and write
permissions for the workflow
* Upate the three tf- shared workflows to adjust the GITHUB_TOKEN
permissions to "read-all"

Side effects of this change:
None.