New ECR Workflows
Why these changes are being introduced:
InfraEng is moving forward with the a plan to allow developers to build
and deploy containers using either X86_64 or ARM64 CPU architecture.
The current ECR publishing workflows in this repo do not support this
option. At the same time, we are doing a little bit of extra
maintenance on the repository by introducing the CODEOWNERS file to
help automate tagging reviewers when workflows are updated.

How this addresses that need:
* Create a CODEOWNERS file and assign files/folders to different teams
in the GitHub Organization
* Create three new ECR publishing workflows that support the option to
choose the CPU architecture for a containers
* Update the README

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1445