Try picking correct runner
Why these changes are being introduced:
The default for GitHub hosted runners is ubuntu on amd64. This caused
the arm64 build to take over 4 minutes (while the amd64 build took only
30 seconds). There are arm64 hosts, so we're trying to pick the runner
based on the CPU_ARCH passed in as an input.

How this implements that change:
* Add a conditional to the runs-on for the job to pick a host based on
the CPU_ARCH value passed in

Related Jira tickets:
* https://mitlibraries.atlassian.net/browse/IN-1445
* https://mitlibraries.atlassian.net/browse/IN-1446