Move Away From Docker BuildX
Why these changes are being introduced:
After experimenting with docker buildx and experimenting with choosing
the hosted runner platform before starting the job, the docker buildx
lost out due to time and complexity. In the end, it was simpler to
make a choice for the cpu architecture of the hosted runner first, and
then just use regular docker build to build for the same platform.

After verifying that, there were a number of minor modifications to
improve the report summaries in the GitHub Actions UI for dev/stage/prod
workflows.

How these changes are being implemented:
* Remove the buildx setup step from the dev and stage workflows
* Switch from the Docker action back to a bash shell to just run the
regular docker commands to build and push the container
* Update the tagging to support the current use of the "latest" tag
while preparing for future tags that look like `latest-amd64` and
`latest-arm64`
* Add the CPU_ARCH input so that we are prepared for future versions
of the "latest" tags (see bullet point above)
* Add back the actions/checkout for the prod workflow so that we can
successfully check for the existence of the .aws-architecture file in
the repository
* Clean up the logic in the re-tag and push step to make it clearer
what is happening
* Create GHA Summary outputs to make it easier for users to see what
happened during a build job
* Update the README to reflect the latest changes

Related Jira Tickets:
* https://mitlibraries.atlassian.net/browse/IN-1445